+++ src/commands/run.rs
@@ -7,13 +7,13 @@ use std::time::Duration;
 use std::{
     ffi::{OsStr, OsString},
     fs::File,
-    path::{Component, Path, PathBuf},
+    path::{Component, PathBuf},
     process,
 };
 use structopt::{clap::AppSettings, StructOpt};
 use wasi_common::preopen_dir;
-use wasmtime::{Engine, Instance, Module, Store, Trap, Val, ValType};
-use wasmtime_wasi::{old::snapshot_0::Wasi as WasiSnapshot0, Wasi};
+use wasmtime::{Engine, Instance, Linker, Module, Store, Trap, Val, ValType};
+use wasmtime_wasi::wasi_linker;
 
 fn parse_module(s: &OsStr) -> Result<PathBuf, OsString> {
     // Do not accept wasmtime subcommand names as the module name
@@ -49,6 +49,14 @@ fn parse_dur(s: &str) -> Result<Duration> {
     // ... otherwise try to parse it with units such as `3s` or `300ms`
     let dur = humantime::parse_duration(s)?;
     Ok(dur)
+}
+
+fn parse_preloads(s: &str) -> Result<(String, PathBuf)> {
+    let parts: Vec<&str> = s.splitn(2, '=').collect();
+    if parts.len() != 2 {
+        bail!("must contain exactly one equals character ('=')");
+    }
+    Ok((parts[0].into(), parts[1].into()))
 }
 
 /// Runs a WebAssembly module
@@ -87,10 +95,10 @@ pub struct RunCommand {
     #[structopt(
         long = "preload",
         number_of_values = 1,
-        value_name = "MODULE_PATH",
-        parse(from_os_str)
+        value_name = "NAME::MODULE_PATH",
+        parse(try_from_str = parse_preloads)
     )]
-    preloads: Vec<PathBuf>,
+    preloads: Vec<(String, PathBuf)>,
 
     /// Maximum execution time of wasm code before timing out (1, 2s, 100ms, etc)
     #[structopt(
@@ -127,17 +135,35 @@ impl RunCommand {
         let preopen_dirs = self.compute_preopen_dirs()?;
         let argv = self.compute_argv();
 
-        let module_registry = ModuleRegistry::new(&store, &preopen_dirs, &argv, &self.vars)?;
+        let mut linker = wasi_linker(&store, &preopen_dirs, &argv, &self.vars)?;
 
         // Load the preload wasm modules.
-        for preload in self.preloads.iter() {
-            Self::instantiate_module(&store, &module_registry, preload)
-                .with_context(|| format!("failed to process preload at `{}`", preload.display()))?;
+        for (name, path) in self.preloads.iter() {
+            // Read the wasm module binary either as `*.wat` or a raw binary
+            let module = Module::from_file(linker.store(), path)?;
+            let instance = linker
+                .instantiate_wasi_abi(&module)
+                .context(format!("failed to instantiate {:?}", path))?;
+
+            // If it was a command, don't register it.
+            let instance = if let Some(instance) = instance {
+                instance
+            } else {
+                continue;
+            };
+
+            linker.instance(name, &instance).with_context(|| {
+                format!(
+                    "failed to process preload `{}` at `{}`",
+                    name,
+                    path.display()
+                )
+            })?;
         }
 
         // Load the main wasm module.
         match self
-            .handle_module(&store, &module_registry)
+            .load_main_module(&mut linker)
             .with_context(|| format!("failed to run main module `{}`", self.module.display()))
         {
             Ok(()) => (),
@@ -220,67 +246,31 @@ impl RunCommand {
         result
     }
 
-    fn instantiate_module(
-        store: &Store,
-        module_registry: &ModuleRegistry,
-        path: &Path,
-    ) -> Result<Instance> {
-        // Read the wasm module binary either as `*.wat` or a raw binary
-        let data = wat::parse_file(path)?;
-
-        let module = Module::new(store, &data)?;
-
-        // Resolve import using module_registry.
-        let imports = module
-            .imports()
-            .map(|i| {
-                let export = match i.module() {
-                    "wasi_snapshot_preview1" => {
-                        module_registry.wasi_snapshot_preview1.get_export(i.name())
-                    }
-                    "wasi_unstable" => module_registry.wasi_unstable.get_export(i.name()),
-                    other => bail!("import module `{}` was not found", other),
-                };
-                match export {
-                    Some(export) => Ok(export.clone().into()),
-                    None => bail!(
-                        "import `{}` was not found in module `{}`",
-                        i.name(),
-                        i.module()
-                    ),
-                }
-            })
-            .collect::<Result<Vec<_>, _>>()?;
-
-        let instance = Instance::new(&module, &imports)
-            .context(format!("failed to instantiate {:?}", path))?;
-
-        Ok(instance)
-    }
-
-    fn handle_module(&self, store: &Store, module_registry: &ModuleRegistry) -> Result<()> {
+    fn load_main_module(&self, linker: &mut Linker) -> Result<()> {
         if let Some(timeout) = self.wasm_timeout {
-            let handle = store.interrupt_handle()?;
+            let handle = linker.store().interrupt_handle()?;
             thread::spawn(move || {
                 thread::sleep(timeout);
                 handle.interrupt();
             });
         }
-        let instance = Self::instantiate_module(store, module_registry, &self.module)?;
+
+        // Read the wasm module binary either as `*.wat` or a raw binary
+        let module = Module::from_file(linker.store(), &self.module)?;
+        let instance = linker
+            .instantiate_wasi_abi(&module)
+            .context(format!("failed to instantiate {:?}", self.module))?;
 
         // If a function to invoke was given, invoke it.
         if let Some(name) = self.invoke.as_ref() {
-            self.invoke_export(instance, name)?;
-        } else if instance.exports().any(|export| export.name().is_empty()) {
-            // Launch the default command export.
-            self.invoke_export(instance, "")?;
+            if let Some(instance) = instance {
+                self.invoke_export(instance, name)
+            } else {
+                bail!("Cannot invoke exports on a command after it has executed")
+            }
         } else {
-            // If the module doesn't have a default command export, launch the
-            // _start function if one is present, as a compatibility measure.
-            self.invoke_export(instance, "_start")?;
+            Ok(())
         }
-
-        Ok(())
     }
 
     fn invoke_export(&self, instance: Instance, name: &str) -> Result<()> {
@@ -344,44 +334,5 @@ impl RunCommand {
         }
 
         Ok(())
-    }
-}
-
-struct ModuleRegistry {
-    wasi_snapshot_preview1: Wasi,
-    wasi_unstable: WasiSnapshot0,
-}
-
-impl ModuleRegistry {
-    fn new(
-        store: &Store,
-        preopen_dirs: &[(String, File)],
-        argv: &[String],
-        vars: &[(String, String)],
-    ) -> Result<ModuleRegistry> {
-        let mut cx1 = wasi_common::WasiCtxBuilder::new();
-
-        cx1.inherit_stdio().args(argv).envs(vars);
-
-        for (name, file) in preopen_dirs {
-            cx1.preopened_dir(file.try_clone()?, name);
-        }
-
-        let cx1 = cx1.build()?;
-
-        let mut cx2 = wasi_common::old::snapshot_0::WasiCtxBuilder::new();
-
-        cx2.inherit_stdio().args(argv).envs(vars);
-
-        for (name, file) in preopen_dirs {
-            cx2.preopened_dir(file.try_clone()?, name);
-        }
-
-        let cx2 = cx2.build()?;
-
-        Ok(ModuleRegistry {
-            wasi_snapshot_preview1: Wasi::new(store, cx1),
-            wasi_unstable: WasiSnapshot0::new(store, cx2),
-        })
     }
 }